Visualize the rise of COVID-19 cases globally
Within months, COVID-19 went from an epidemic to a pandemic. From the first identified case in December 2019, how did the virus spread so fast and widely? In this project, we will visualize data from the early months of the coronavirus outbreak to see how this virus grew to be a global pandemic.